Implemented the final report on the Socket server. New field on the app, that refers to the csv file name.
executing the Main() of the Socket server, with option -m server, it will continue acting as a server.
when use the option -m report, it will get all csv file inside of it's folder, and get the average of jitter, packet loss and error lost, all between the eschange
of one ap to another.